Move testIfAnyDisputeTriedReplay and related methods to DisputeValidation
