+++ core/src/main/java/bisq/core/trade/bisq_v1/TradeDataValidation.java
@@ -20,12 +20,9 @@ package bisq.core.trade.bisq_v1;
 import bisq.core.btc.wallet.BtcWalletService;
 import bisq.core.dao.DaoFacade;
 import bisq.core.offer.Offer;
-import bisq.core.support.SupportType;
 import bisq.core.support.dispute.Dispute;
 import bisq.core.support.dispute.DisputeValidation;
 import bisq.core.trade.model.bisq_v1.Trade;
-
-import bisq.common.util.Tuple3;
 
 import org.bitcoinj.core.Address;
 import org.bitcoinj.core.Coin;
@@ -35,11 +32,6 @@ import org.bitcoinj.core.TransactionInput;
 import org.bitcoinj.core.TransactionOutPoint;
 import org.bitcoinj.core.TransactionOutput;
 
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
 import java.util.function.Consumer;
 
 import lombok.Getter;
@@ -52,125 +44,6 @@ import static com.google.common.base.Preconditions.checkNotNull;
 
 @Slf4j
 public class TradeDataValidation {
-
-    public static void testIfAnyDisputeTriedReplay(List<Dispute> disputeList,
-                                                   Consumer<DisputeValidation.DisputeReplayException> exceptionHandler) {
-        var tuple = getTestReplayHashMaps(disputeList);
-        Map<String, Set<String>> disputesPerTradeId = tuple.first;
-        Map<String, Set<String>> disputesPerDelayedPayoutTxId = tuple.second;
-        Map<String, Set<String>> disputesPerDepositTxId = tuple.third;
-
-        disputeList.forEach(disputeToTest -> {
-            try {
-                testIfDisputeTriesReplay(disputeToTest,
-                        disputesPerTradeId,
-                        disputesPerDelayedPayoutTxId,
-                        disputesPerDepositTxId);
-
-            } catch (DisputeValidation.DisputeReplayException e) {
-                exceptionHandler.accept(e);
-            }
-        });
-    }
-
-
-    public static void testIfDisputeTriesReplay(Dispute dispute,
-                                                List<Dispute> disputeList) throws DisputeValidation.DisputeReplayException {
-        var tuple = TradeDataValidation.getTestReplayHashMaps(disputeList);
-        Map<String, Set<String>> disputesPerTradeId = tuple.first;
-        Map<String, Set<String>> disputesPerDelayedPayoutTxId = tuple.second;
-        Map<String, Set<String>> disputesPerDepositTxId = tuple.third;
-
-        testIfDisputeTriesReplay(dispute,
-                disputesPerTradeId,
-                disputesPerDelayedPayoutTxId,
-                disputesPerDepositTxId);
-    }
-
-
-    private static Tuple3<Map<String, Set<String>>, Map<String, Set<String>>, Map<String, Set<String>>> getTestReplayHashMaps(
-            List<Dispute> disputeList) {
-        Map<String, Set<String>> disputesPerTradeId = new HashMap<>();
-        Map<String, Set<String>> disputesPerDelayedPayoutTxId = new HashMap<>();
-        Map<String, Set<String>> disputesPerDepositTxId = new HashMap<>();
-        disputeList.forEach(dispute -> {
-            String uid = dispute.getUid();
-
-            String tradeId = dispute.getTradeId();
-            disputesPerTradeId.putIfAbsent(tradeId, new HashSet<>());
-            Set<String> set = disputesPerTradeId.get(tradeId);
-            set.add(uid);
-
-            String delayedPayoutTxId = dispute.getDelayedPayoutTxId();
-            if (delayedPayoutTxId != null) {
-                disputesPerDelayedPayoutTxId.putIfAbsent(delayedPayoutTxId, new HashSet<>());
-                set = disputesPerDelayedPayoutTxId.get(delayedPayoutTxId);
-                set.add(uid);
-            }
-
-            String depositTxId = dispute.getDepositTxId();
-            if (depositTxId != null) {
-                disputesPerDepositTxId.putIfAbsent(depositTxId, new HashSet<>());
-                set = disputesPerDepositTxId.get(depositTxId);
-                set.add(uid);
-            }
-        });
-
-        return new Tuple3<>(disputesPerTradeId, disputesPerDelayedPayoutTxId, disputesPerDepositTxId);
-    }
-
-    private static void testIfDisputeTriesReplay(Dispute disputeToTest,
-                                                 Map<String, Set<String>> disputesPerTradeId,
-                                                 Map<String, Set<String>> disputesPerDelayedPayoutTxId,
-                                                 Map<String, Set<String>> disputesPerDepositTxId)
-            throws DisputeValidation.DisputeReplayException {
-
-        try {
-            String disputeToTestTradeId = disputeToTest.getTradeId();
-            String disputeToTestDelayedPayoutTxId = disputeToTest.getDelayedPayoutTxId();
-            String disputeToTestDepositTxId = disputeToTest.getDepositTxId();
-            String disputeToTestUid = disputeToTest.getUid();
-
-            // For pre v1.4.0 we do not get the delayed payout tx sent in mediation cases but in refund agent case we do.
-            // So until all users have updated to 1.4.0 we only check in refund agent case. With 1.4.0 we send the
-            // delayed payout tx also in mediation cases and that if check can be removed.
-            if (disputeToTest.getSupportType() == SupportType.REFUND) {
-                checkNotNull(disputeToTestDelayedPayoutTxId,
-                        "Delayed payout transaction ID is null. " +
-                                "Trade ID: " + disputeToTestTradeId);
-            }
-            checkNotNull(disputeToTestDepositTxId,
-                    "depositTxId must not be null. Trade ID: " + disputeToTestTradeId);
-            checkNotNull(disputeToTestUid,
-                    "agentsUid must not be null. Trade ID: " + disputeToTestTradeId);
-
-            Set<String> disputesPerTradeIdItems = disputesPerTradeId.get(disputeToTestTradeId);
-            checkArgument(disputesPerTradeIdItems != null && disputesPerTradeIdItems.size() <= 2,
-                    "We found more then 2 disputes with the same trade ID. " +
-                            "Trade ID: " + disputeToTestTradeId);
-            if (!disputesPerDelayedPayoutTxId.isEmpty()) {
-                Set<String> disputesPerDelayedPayoutTxIdItems = disputesPerDelayedPayoutTxId.get(disputeToTestDelayedPayoutTxId);
-                checkArgument(disputesPerDelayedPayoutTxIdItems != null && disputesPerDelayedPayoutTxIdItems.size() <= 2,
-                        "We found more then 2 disputes with the same delayedPayoutTxId. " +
-                                "Trade ID: " + disputeToTestTradeId);
-            }
-            if (!disputesPerDepositTxId.isEmpty()) {
-                Set<String> disputesPerDepositTxIdItems = disputesPerDepositTxId.get(disputeToTestDepositTxId);
-                checkArgument(disputesPerDepositTxIdItems != null && disputesPerDepositTxIdItems.size() <= 2,
-                        "We found more then 2 disputes with the same depositTxId. " +
-                                "Trade ID: " + disputeToTestTradeId);
-            }
-        } catch (IllegalArgumentException e) {
-            throw new DisputeValidation.DisputeReplayException(disputeToTest, e.getMessage());
-        } catch (NullPointerException e) {
-            log.error("NullPointerException at testIfDisputeTriesReplay: " +
-                            "disputeToTest={}, disputesPerTradeId={}, disputesPerDelayedPayoutTxId={}, " +
-                            "disputesPerDepositTxId={}",
-                    disputeToTest, disputesPerTradeId, disputesPerDelayedPayoutTxId, disputesPerDepositTxId);
-            throw new DisputeValidation.DisputeReplayException(disputeToTest, e.toString() + " at dispute " + disputeToTest.toString());
-        }
-    }
-
     public static void validateDelayedPayoutTx(Trade trade,
                                                Transaction delayedPayoutTx,
                                                DaoFacade daoFacade,
